+++ sdk/core/Azure.Core/tests/Azure.Core.Tests.csproj
@@ -49,6 +49,7 @@
     <Compile Include="..\..\Azure.Core.TestFramework\src\Instrumentation\ZeroPollingStrategy.cs" LinkBase="Shared\TestFramework" />
     <Compile Include="$(TestFrameworkSupportFiles)" LinkBase="Shared\TestFramework" />
     <Compile Include="..\src\Shared\ApiVersionString.cs" LinkBase="Shared" />
+    <Compile Include="..\src\Shared\GuidUtilities.cs" LinkBase="Shared" />
   </ItemGroup>
   <ItemGroup>
     <Compile Update="PemReaderTests.Keys.cs" DependentUpon="PemReaderTests.cs" />
